// src/types/vehicle.ts
export type VehicleRow = {
  v_id: number | string;
  v_name?: string;
  v_route?: string | number;
  v_class?: string | number;
  v_metr?: string | number;
  v_plate?: string;
  v_chassis?: string;
  v_engine?: string;
  v_brand?: string;
  v_model?: string;
  v_company?: string;
  v_inv_com?: string;
  v_register?: string;
  v_note?: string;
};